% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCIndep.R
\name{ASCIndep}
\alias{ASCIndep}
\title{Asymptotic Score Continuity-corrected Confidence Interval for Independent RR}
\usage{
ASCIndep(x, m = NULL, level = 0.95, tol = 1e-10)
}
\arguments{
\item{x}{A four-element vector \code{c(x11, x12, x21, x22)} represents the numbers
in a 2-by-2 contingency table. See details.}

\item{m}{An number indicates the strength of the continuity correction. The
default is \code{NULL}, which means no continuity correction
\insertCite{miettinen1985comparative}{ASCCIRR}. We recommend
\code{2} if the continuity correction is needed.}

\item{level}{Confidence level. Default is \code{0.95}.}

\item{tol}{The desired accuracy (convergence tolerance) for the iterative root
finding procedure when finding intevals. Default is \code{1e-10}.}
}
\value{
A list of two objects: \code{result} and \code{method}.
}
\description{
Asymptotic Score Continuity-corrected Confidence Interval for Independent RR
}
\details{
The contingency table is considered as:\tabular{lcr}{
    \tab Event \tab Non-Event \cr
   Group 1 \tab x11 \tab x12 \cr
   Group 2 \tab x21 \tab x22 \cr
}
}
\examples{
x <- c(7, 27, 1, 33)

ASCIndep(x)
ASCIndep(x, m=2)   # ASC

}
\references{
{
\insertRef{miettinen1985comparative}{ASCCIRR}
}
}
