% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCPaired.R
\name{ASCPaired}
\alias{ASCPaired}
\title{Asymptotic Score Continuity-corrected Confidence Interval for Paired RR}
\usage{
ASCPaired(x, m = NULL, level = 0.95, ctrl = 1e-10)
}
\arguments{
\item{x}{A four-element vector \code{c(x11, x12, x21, x22)} represents the numbers
in a 2-by-2 contingency table. See details.}

\item{m}{An number indicates the strength of the continuity correction. The
default is \code{NULL}, which means no continuity correction
\insertCite{tang2012confidence}{ASCCIRR}. And the values
of \code{2}, \code{4}, and \code{8} represent the continuity corrections of high,
medium, and low strength, respectively \insertCite{delrocco2023new}{ASCCIRR}.}

\item{level}{Confidence level. Default is \code{0.95}.}

\item{ctrl}{A very small number will be added to the contingency table if \code{0} exists.
Default is \code{1e-10}.}
}
\value{
A list of two objects: \code{result} and \code{method}.
}
\description{
Asymptotic Score Continuity-corrected Confidence Interval for Paired RR
}
\details{
The contingency table is considered as:\tabular{lcr}{
    \tab Test2: Event \tab Test2: Nonevent \cr
   Test1: Event \tab x11 \tab x12 \cr
   Test1: Nonevent \tab x21 \tab x22 \cr
}
}
\examples{
x <- c(9, 36, 14, 1371)

ASCPaired(x)
ASCPaired(x, m=2)  # ASCC-H
ASCPaired(x, m=4)  # ASCC-M
ASCPaired(x, m=8)  # ASCC-L

}
\references{
{
\insertRef{tang2012confidence}{ASCCIRR}
\insertRef{delrocco2023new}{ASCCIRR}
}
}
